feat(camera): send where the motion has been detected
Where is the Consideration object. The main app sends us some ROI (or not), then we create these ROI with the Consideration data: what type is it (rectangle roi? all?) with the id. Thanks to this, the main app can know where the motion has been detected: which ROI (it can retrieve it thanks to the id and type to know the database table for ex), or if it's on the whole camera because it didn't define any ROI, meaning that the ROI is basically the whole camera view.