feat(core): mqtt bool payload parse accepts string and byte (#239)
Before this it was only accepting 0x00 (false) and 0x01 (true).
I added this because I did not find any solution to send these bytes for will message with the PubSub library used by the espcam.
It adds complexity that I don't wanted but its well tested and as of today I don't have other choice, will message is necessary to know either or not the esp is still responding. Thus it adds flexibility that could be neat.