feat(core): turn off alarm integration with services.
It also controls either or not we should sync services with the alarm status. If a motion is being detected, we should not say to services that the alarm turned off because the alarm service is not actually off. We can force the service to turned off manually (i.e user action).