wip fix(smart-camera): exit main process when camera process stop.
We have the main process (Manager) which turns on/off the camera with 2 params: stream (bool) and to_analyze (bool). The on/off works by running or killing a separate python Process. But when this process crashes, the main process does not know about it and continue its life as nothing happened BUT the camera is not working anymore (it does not deliver frames). The end user receives a notification about it but he needs to restart the docker service because everything is still running.
One solution : when the process exit anormally (not a SIGTERM sent by terminate()), then we stop the main process so docker will restart the service to fix the issue.