feat(camera): detect people for x seconds before triggering motion event (#261)
To trigger detection, the system needs to detect at least someone for X consecutive seconds.

Explanation:
Sometimes we get false positives, meaning it will detect a human even if nobody is present. It's an edge case but it happens in noisy environments. The idea behind this PR is to add something like a threshold well known in electronics. We don't wan't to trigger "motion" / "no more motion" immediately after one detection or non-detection (1 frame), otherwise it leads to a lot of "noise" (imagine trigger/stop your heating around a fixed value without any delta...). For the non-detection it's already done, with this PR we do the detection.